Updating pytorch GAN implementation after merging; consolidated Vocab object implementations; updated import paths; re-implemented sample generation and pre-training for generator.

+++ src/gan/discriminator/treelstm/utils.py
@@ -26,7 +26,7 @@ def load_word_vectors(path):
         with open(path, "rb") as f:
             glove = pickle.load(f)
 
-        vectors = torch.from_numpy(glove['embeddings'])
+        vectors = torch.from_numpy(glove['embeddings']).float()
         vocab = Vocab(data=glove['vocab'])
 
         return vocab, vectors
